--- app/(layout)/(components)/(tap-system)/tab-open-utils.ts
@@ -1,7 +1,0 @@
-export const tabOpen = (href: string): void => {
-  const link = document.createElement('a')
-  link.href = href
-  document.body.appendChild(link)
-  link.click()
-  document.body.removeChild(link)
-}
