+++ app/(layout)/(components)/(tap-system)/tap-bar-main.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react'
-import { TabBar } from './TabBar'
+import { TabBar } from './tab-bar'
 import { usePathname } from 'next/navigation'
-import { useTabs } from '../../(usecase)/TabUseCases'
+import { useTabs } from '../../(usecase)/tab-usecases'
 
 export function TapBarMain (): React.ReactElement {
   const path = usePathname()
