+++ app/memo/(components)/memo-system-navigator/FolderAndMemoStructure.tsx
@@ -1,9 +1,9 @@
 'use client'
 import React, { useEffect, useRef, useState } from 'react'
 import FolderItem from './FolderItem'
-import TabOpen from '../../../(layout)/(components)/(tap-system)/TabOpen'
-import MemoItem from './MemoItem'
-import { type ContextMenuProps } from './MemoAndFolderContextMenu'
+import TabOpen from '../../../(layout)/(components)/(tap-system)/tab-open'
+import MemoItem from './memo-item'
+import { type ContextMenuProps } from './memo-folder-context-menu'
 import { ApplicationType } from '../../../(layout)/(domain)/tab'
 import { useMemoSystem } from '../../(usecase)/memo-system-usecases'
 import { extractFolderIdByMemoId, type Folder } from '../../(domain)/folder'
