+++ app/memo/(components)/memo-system-navigator/FolderItem.tsx
@@ -4,7 +4,7 @@ import openFolder from '../../../../public/openFolder.png'
 import folderWithContentImg from '../../../../public/contentFolder.png'
 import Image from 'next/image'
 import React, { useEffect, useRef, useState } from 'react'
-import { type ContextMenuProps } from './MemoAndFolderContextMenu'
+import { type ContextMenuProps } from './memo-folder-context-menu'
 import { useFolderAndMemo } from '../../(usecase)/memo-folder-usecases'
 
 export default function FolderItem ({
