+++ app/memo/(components)/memo-system-navigator/memo-folder-context-menu.tsx
@@ -8,7 +8,7 @@ export interface ContextMenuProps {
   folderName?: string
 }
 
-export default function MemoAndFolderContextMenu ({
+export default function MemoFolderContextMenu ({
   contextMenu,
   closeContextMenu,
   handleDeleteClick,
