+++ app/memo/(components)/memo-system-navigator/memo-item.tsx
@@ -2,7 +2,7 @@ import { type SimpleMemoInfo } from '../../(domain)/memo'
 import React, { useState } from 'react'
 import Image from 'next/image'
 import memoImg from '../../../../public/memo.png'
-import { type ContextMenuProps } from './MemoAndFolderContextMenu'
+import { type ContextMenuProps } from './memo-folder-context-menu'
 import { useFolderAndMemo } from '../../(usecase)/memo-folder-usecases'
 import { useMemoSystem } from '../../(usecase)/memo-system-usecases'
